hrt: implement semi-working version of stumpwm's grab-pointer
Change the cursor image when the keybinding state is active, and reset
it once it finishes.

This isn't actually the right solution, as we need to unfocus whatever
is on the seat to avoid the subsequent key presses from being passed
to clients. This would also keep wlroots from changing the cursor
whenever it is moved over a surface.
