+++ heart/include/hrt/hrt_input.h
@@ -89,7 +89,8 @@ struct hrt_input {
 
 /**
  * Send the appropriate cursor event to the view under the
- * cursor.
+ * cursor and ensure the cursor has the correct image
+ * for what it is hovering over.
  **/
 void hrt_seat_reset_view_under(struct hrt_seat *seat);
 
@@ -102,6 +103,11 @@ void hrt_seat_reset_view_under(struct hrt_seat *seat);
  * names.
  */
 void hrt_seat_set_cursor_img(struct hrt_seat *seat, char *img_name);
+
+/**
+ * Set the cursor image back to the default.
+ **/
+void hrt_seat_reset_cursor_img(struct hrt_seat *seat);
 
 void hrt_seat_notify_button(struct hrt_seat *seat,
                             struct wlr_pointer_button_event *event);
