+++ lisp/heart/hrt-bindings.lisp
@@ -57,7 +57,8 @@
 (declaim (inline hrt-seat-reset-view-under))
 (cffi:defcfun ("hrt_seat_reset_view_under" hrt-seat-reset-view-under) :void
   "Send the appropriate cursor event to the view under the
-cursor."
+cursor and ensure the cursor has the correct image
+for what it is hovering over."
   (seat (:pointer (:struct hrt-seat))))
 
 #-HRT-DEBUG
@@ -71,6 +72,12 @@ See themes section of man xcursor(3) to find where to find valid cursor
 names."
   (seat (:pointer (:struct hrt-seat)))
   (img-name (:pointer :char)))
+
+#-HRT-DEBUG
+(declaim (inline hrt-seat-reset-cursor-img))
+(cffi:defcfun ("hrt_seat_reset_cursor_img" hrt-seat-reset-cursor-img) :void
+  "Set the cursor image back to the default."
+  (seat (:pointer (:struct hrt-seat))))
 
 #-HRT-DEBUG
 (declaim (inline hrt-seat-notify-button))
